undetected chromedriver and receiver column added
Exchanged the conventional selenium for a undetected version, so i can handle easily some captchas. Also included a human-like typing function in order to lessen the captcha activation risk. Also, added a receiver column to the front end so the user can modify from which excel column the emails are.